Refactored Services tests
