+++ src/Hero6.Tests.HelperTools/Services/ServiceTestBase.cs
@@ -16,6 +16,7 @@ namespace LateStartStudio.Hero6.Services
         [SetUp]
         public override void SetUp()
         {
+            base.SetUp();
             Service = MakeService();
         }
